refactor(client): reorganize user feature directory
- Move user profile and settings pages to `features/user/pages`
- Move user components to `features/user/components`
- Update exports in `features/user/components/index.ts` and `features/user/pages/index.ts`
