--- apps/client/src/features/user/components/forms/index.ts
@@ -1,2 +1,0 @@
-export * from "./update-password-form";
-export * from "./update-user-form";
