--- apps/client/src/features/user/components/forms/update-user-form.tsx
@@ -1,70 +1,0 @@
-"use client";
-
-import { zodResolver } from "@hookform/resolvers/zod";
-import { Mail, User } from "lucide-react";
-import { useForm } from "react-hook-form";
-
-import type { IUpdateUserDto } from "@tubenote/dtos";
-import { updateUserSchema } from "@tubenote/schemas";
-
-import { FormInput } from "@/components/global";
-import { Button, Form } from "@/components/ui";
-
-import { useUpdateUser } from "../../hooks";
-import { useUserStore } from "../../store";
-
-export function UpdateUserForm() {
-  const { currentUser } = useUserStore();
-  const { mutate: updateUser, isPending } = useUpdateUser();
-
-  const form = useForm<IUpdateUserDto>({
-    resolver: zodResolver(updateUserSchema),
-    defaultValues: {
-      username: currentUser?.username ?? "",
-      email: currentUser?.email ?? "",
-    },
-  });
-
-  const handleUpdateProfile = (data: IUpdateUserDto) => {
-    updateUser(data);
-  };
-
-  return (
-    <div className="bg-white rounded-lg shadow-md p-6 mb-8">
-      <h2 className="text-2xl font-semibold mb-4 flex items-center">
-        <User className="mr-2" />
-        Profile Information
-      </h2>
-      <Form {...form}>
-        <form
-          onSubmit={form.handleSubmit(handleUpdateProfile)}
-          className="space-y-4"
-        >
-          <FormInput
-            name="username"
-            label="Username"
-            placeholder="Your Username"
-            icon={User}
-            control={form.control}
-          />
-          <FormInput
-            name="email"
-            type="email"
-            label="Email"
-            placeholder="your.email@example.com"
-            icon={Mail}
-            control={form.control}
-          />
-          <Button
-            type="submit"
-            className="w-full bg-gradient-to-r from-red-600 to-purple-600 text-white
-						hover:from-red-700 hover:to-purple-700 disabled:opacity-50 disabled:cursor-not-allowed"
-            disabled={isPending}
-          >
-            {isPending ? "Updating..." : "Update Profile"}
-          </Button>
-        </form>
-      </Form>
-    </div>
-  );
-}
