+++ apps/client/src/features/user/components/index.ts
@@ -1,2 +1,4 @@
-export * from "./forms";
-export * from "./ui";
+export * from "./account-actions";
+export * from "./user-avatar";
+export * from "./user-info";
+export * from "./user-profile";
