--- apps/client/src/features/user/components/ui/index.ts
@@ -1,4 +1,0 @@
-export * from "./account-actions";
-export * from "./user-info";
-export * from "./user-avatar";
-export * from "./user-profile";
