+++ apps/client/src/features/user/components/user-avatar.tsx
@@ -2,15 +2,15 @@ import type { User } from "@tubenote/db";
 
 import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
 
-type UserAvatarProps = {
+interface UserAvatarProps {
   user: User;
-};
+}
 
 export function UserAvatar({ user }: UserAvatarProps) {
   const src = "https://github.com/shadcn.png";
 
-  const avatarFallback =
-    user.username[0].toUpperCase() + user.username[1].toUpperCase();
+  const avatarFallback
+    = user.username[0].toUpperCase() + user.username[1].toUpperCase();
 
   const imgSrc = user.profilePicture ?? src;
