+++ apps/client/src/features/user/components/user-profile.tsx
@@ -1,0 +1,32 @@
+import { useUserStore } from "../store";
+import { UserAvatar } from "./";
+
+interface UserProfileProps {
+  isOpen: boolean;
+}
+
+export function UserProfile({ isOpen }: UserProfileProps) {
+  const { currentUser } = useUserStore();
+
+  return (
+    <div className="p-4">
+      <div className="flex flex-col items-center justify-center gap-2 text-center">
+        {currentUser
+          ? (
+              <>
+                <UserAvatar user={currentUser} />
+                {isOpen && (
+                  <div className="hidden md:block">
+                    <h3 className="font-semibold text-gray-700">
+                      {currentUser.username}
+                    </h3>
+                    <p className="text-sm text-gray-500">{currentUser.email}</p>
+                  </div>
+                )}
+              </>
+            )
+          : null}
+      </div>
+    </div>
+  );
+}
