+++ apps/client/src/features/user/pages/user-profile/components/skeletons/user-learning-progress-skeleton.tsx
