+++ apps/client/src/features/user/pages/user-profile/components/skeletons/user-recent-activity-skeleton.tsx
