+++ apps/client/src/features/user/pages/user-profile/components/user-profile-avatar.tsx
