+++ apps/client/src/features/user/pages/user-profile/index.ts
@@ -1,0 +1,1 @@
+export * from "./page";
