+++ apps/client/src/features/user/pages/user-settings/index.ts
@@ -1,0 +1,1 @@
+export * from "./page";
